+++ src/Operation/Internal/Operations.php
@@ -22,8 +22,8 @@ use FiiSoft\Jackdaw\Operation\Operation;
 use FiiSoft\Jackdaw\Operation\Sending\{CollectIn, CollectKeysIn, CountIn, Dispatch, Feed, FeedMany, Remember, RouteMany,
     RouteOne, SendTo, SendToMany, SendToMax, SendWhen, SendWhileUntil\SendUntil, SendWhileUntil\SendWhile, StoreIn,
     Unzip};
-use FiiSoft\Jackdaw\Operation\Special\{Assert, CollectInArray, Limit, ReadMany, ReadNext, ReadWhileUntil\ReadUntil,
-    ReadWhileUntil\ReadWhile, Shuffle, WhileUntil\UntilTrue, WhileUntil\WhileTrue};
+use FiiSoft\Jackdaw\Operation\Special\{Assert, CollectInArray, IterateOver, Limit, ReadMany, ReadNext,
+    ReadWhileUntil\ReadUntil, ReadWhileUntil\ReadWhile, Shuffle, WhileUntil\UntilTrue, WhileUntil\WhileTrue};
 use FiiSoft\Jackdaw\Operation\Terminating\{Collect, CollectKeys, Count, FinalOperation, Find, First, Fold, GroupBy, Has,
     HasEvery, HasOnly, IsEmpty, Last, Reduce};
 use FiiSoft\Jackdaw\Producer\ProducerReady;
@@ -385,7 +385,7 @@ final class Operations
     /**
      * @param DiscriminatorReady|callable|array<string|int>|string|int $discriminator
      */
-    public static function chunkBy($discriminator, bool $reindex = false): Operation
+    public static function chunkBy($discriminator, ?bool $reindex = null): Operation
     {
         return ChunkBy::create($discriminator, $reindex);
     }
@@ -437,6 +437,11 @@ final class Operations
     public static function flat(int $level = 0): Flat
     {
         return new Flat($level);
+    }
+    
+    public static function iterateOver(callable $producer): IterateOver
+    {
+        return IterateOver::create($producer);
     }
     
     /**