+++ src/Operation/Terminating/GroupBy.php
@@ -4,6 +4,7 @@ namespace FiiSoft\Jackdaw\Operation\Terminating;
 
 use FiiSoft\Jackdaw\Discriminator\DiscriminatorReady;
 use FiiSoft\Jackdaw\Internal\Collection\BaseStreamCollection;
+use FiiSoft\Jackdaw\Internal\Helper;
 use FiiSoft\Jackdaw\Operation\Internal\GroupingOperation;
 use FiiSoft\Jackdaw\Operation\Terminating\GroupBy\GroupByReindexKeys;
 
@@ -14,7 +15,7 @@ abstract class GroupBy extends GroupingOperation
      */
     final public static function create($discriminator, ?bool $reindex = null): self
     {
-        return self::shouldReindex($discriminator, $reindex)
+        return Helper::shouldReindex($discriminator, $reindex)
             ? new GroupByReindexKeys($discriminator)
             : new GroupBy\GroupByKeepKeys($discriminator);
     }