+++ tests/StreamHTest.php
@@ -7,6 +7,7 @@ use FiiSoft\Jackdaw\Discriminator\Discriminators;
 use FiiSoft\Jackdaw\Handler\OnError;
 use FiiSoft\Jackdaw\Internal\Check;
 use FiiSoft\Jackdaw\Mapper\Mappers;
+use FiiSoft\Jackdaw\Memo\FullMemo;
 use FiiSoft\Jackdaw\Memo\Memo;
 use FiiSoft\Jackdaw\Operation\Collecting\Fork\Adapter\IdleForkHandler;
 use FiiSoft\Jackdaw\Operation\Exception\OperationExceptionFactory;
@@ -609,5 +610,196 @@ final class StreamHTest extends TestCase
         $this->expectExceptionObject(AssertionFailed::exception(null, 2, Check::VALUE));
         
         \iterator_to_array(Stream::from(['a', 'b', null, 'c', 'd'])->onError(OnError::skip())->assert('is_string'));
+    }
+    
+    /**
+     * @dataProvider getDataForTestIterateOverZeroArgs
+     */
+    #[DataProvider('getDataForTestIterateOverZeroArgs')]
+    public function test_iterateOver_zero_args(FullMemo $num, callable $producer): void
+    {
+        $actual = Stream::from([2, 0, 1, 3, 1, 2])
+            ->remember($num)
+            ->iterate($producer)
+            ->without(['e', 'y', 'u', 'i', 'o', 'a'])
+            ->chunkBy($num->key())
+            ->concat('')
+            ->toString();
+        
+        self::assertSame('hp,chj,kqn,bzvr,kqn,hp', $actual);
+    }
+    
+    public static function getDataForTestIterateOverZeroArgs(): iterable
+    {
+        $dataSet = [
+            ['c', 'e', 'h', 'j'],
+            ['k', 'q', 'n'],
+            ['h', 'p'],
+            ['b', 'z', 'o', 'v', 'r'],
+        ];
+        
+        $num = Memo::full();
+        
+        yield 'array' => [$num, static fn(): array => $dataSet[$num->value()->read()]];
+
+        yield 'iterable' => [$num, static fn(): iterable => $dataSet[$num->value()->read()]];
+
+        yield 'Generator' => [$num, static function () use ($dataSet, $num): \Generator {
+            yield from $dataSet[$num->value()->read()];
+        }];
+
+        yield 'Iterator' => [$num, static function () use ($dataSet, $num): \Iterator {
+            return new \ArrayIterator($dataSet[$num->value()->read()]);
+        }];
+        
+        yield 'Traversable' => [$num, static function () use ($dataSet, $num): \Traversable {
+            return new \ArrayIterator($dataSet[$num->value()->read()]);
+        }];
+    }
+    
+    public function test_iterateOver_zero_args_with_onerror_handler(): void
+    {
+        $dataSet = [
+            ['c', 'e', 'h', 'j'],
+            ['k', 'q', 'n'],
+            ['h', 'p'],
+            ['b', 'z', 'o', 'v', 'r'],
+        ];
+        
+        $num = Memo::full();
+        
+        $actual = Stream::from([2, 0, 1, 3, 1, 2])
+            ->onError(OnError::abort())
+            ->remember($num)
+            ->iterate(static fn(): array => $dataSet[$num->value()->read()])
+            ->without(['e', 'y', 'u', 'i', 'o', 'a'])
+            ->chunkBy($num->key())
+            ->concat('')
+            ->toString();
+        
+        self::assertSame('hp,chj,kqn,bzvr,kqn,hp', $actual);
+    }
+    
+    public function test_iterateOver_one_arg(): void
+    {
+        $dataSet = [
+            ['c', 'e', 'h', 'j'],
+            ['k', 'q', 'n'],
+            ['h', 'p'],
+            ['b', 'z', 'o', 'v', 'r'],
+        ];
+        
+        $actual = Stream::from([2, 0, 1, 3, 1, 2])
+            ->putIn($key, Check::KEY)
+            ->iterate(static fn(int $num): iterable => $dataSet[$num])
+            ->without(['e', 'y', 'u', 'i', 'o', 'a'])
+            ->chunkBy(Discriminators::readFrom($key))
+            ->concat('')
+            ->toString();
+        
+        self::assertSame('hp,chj,kqn,bzvr,kqn,hp', $actual);
+    }
+    
+    public function test_iterateOver_one_arg_with_onerror_handler(): void
+    {
+        $dataSet = [
+            ['c', 'e', 'h', 'j'],
+            ['k', 'q', 'n'],
+            ['h', 'p'],
+            ['b', 'z', 'o', 'v', 'r'],
+        ];
+        
+        $actual = Stream::from([2, 0, 1, 3, 1, 2])
+            ->onError(OnError::abort())
+            ->putIn($key, Check::KEY)
+            ->iterate(static fn(int $num): array => $dataSet[$num])
+            ->without(['e', 'y', 'u', 'i', 'o', 'a'])
+            ->chunkBy(Discriminators::readFrom($key))
+            ->concat('')
+            ->toString();
+        
+        self::assertSame('hp,chj,kqn,bzvr,kqn,hp', $actual);
+    }
+    
+    public function test_iterateOver_two_args(): void
+    {
+        $dataSet = [
+            ['c', 'e', 'h', 'j'],
+            ['k', 'q', 'n'],
+            ['h', 'p'],
+            ['b', 'z', 'o', 'v', 'r'],
+        ];
+        
+        $actual = Stream::from([2, 0, 1, 3, 1, 2])
+            ->iterate(static function (int $num, int $key) use ($dataSet): \Generator {
+                foreach ($dataSet[$num] as $value) {
+                    yield $key => $value;
+                }
+            })
+            ->without(['e', 'y', 'u', 'i', 'o', 'a'])
+            ->chunkByKey()
+            ->concat('')
+            ->toString();
+        
+        self::assertSame('hp,chj,kqn,bzvr,kqn,hp', $actual);
+    }
+    
+    public function test_iterateOver_two_args_with_onerror_handler(): void
+    {
+        $dataSet = [
+            ['c', 'e', 'h', 'j'],
+            ['k', 'q', 'n'],
+            ['h', 'p'],
+            ['b', 'z', 'o', 'v', 'r'],
+        ];
+        
+        $actual = Stream::from([2, 0, 1, 3, 1, 2])
+            ->onError(OnError::abort())
+            ->iterate(static function (int $num, int $key) use ($dataSet): \Generator {
+                foreach ($dataSet[$num] as $value) {
+                    yield $key => $value;
+                }
+            })
+            ->without(['e', 'y', 'u', 'i', 'o', 'a'])
+            ->chunkByKey()
+            ->concat('')
+            ->toString();
+        
+        self::assertSame('hp,chj,kqn,bzvr,kqn,hp', $actual);
+    }
+    
+    public function test_use_flatMap_instead_of_iterateOver(): void
+    {
+        $dataSet = [
+            ['c', 'e', 'h', 'j'],
+            ['k', 'q', 'n'],
+            ['h', 'p'],
+            ['b', 'z', 'o', 'v', 'r'],
+        ];
+        
+        $actual = Stream::from([2, 0, 1, 3, 1, 2])
+            ->putIn($key, Check::KEY)
+            ->flatMap(static fn(int $num): array => $dataSet[$num], 1)
+            ->without(['e', 'y', 'u', 'i', 'o', 'a'])
+            ->chunkBy(Discriminators::readFrom($key))
+            ->concat('')
+            ->toString();
+        
+        self::assertSame('hp,chj,kqn,bzvr,kqn,hp', $actual);
+    }
+    
+    public function test_iterateOver_throws_exception_when_callable_requires_invalid_number_of_args(): void
+    {
+        $this->expectExceptionObject(OperationExceptionFactory::invalidIterateOverCallback(3));
+        
+        Stream::empty()->iterate(static fn($a, $b, $c): iterable => []);
+    }
+    
+    
+    public function test_iterateOver_throws_exception_when_declared_type_of_callable_is_invalid(): void
+    {
+        $this->expectExceptionObject(OperationExceptionFactory::wrongTypeOfIterateOverCallback());
+        
+        Stream::empty()->iterate(static fn(): int => 5);
     }
 }